Replaced PlayerAnimationController strings with variables
