--- Assets/Scripts/animation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d708b61159c0bdf4c825273d72cc2b61
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
